+++ src/ts/util/selection.ts
@@ -1,6 +1,6 @@
 import {Constants} from "../constants";
 import {isChrome} from "./compatibility";
-import {hasClosestBlock} from "./hasClosest";
+import {hasClosestBlock, hasClosestByClassName} from "./hasClosest";
 
 export const getEditorRange = (element: HTMLElement) => {
     let range: Range;
@@ -19,7 +19,7 @@ export const getEditorRange = (element: HTMLElement) => {
 
 export const getCursorPosition = (editor: HTMLElement) => {
     const range = window.getSelection().getRangeAt(0);
-    if (!editor.contains(range.startContainer)) {
+    if (!editor.contains(range.startContainer) && !hasClosestByClassName(range.startContainer, "vditor-panel--none")) {
         return {
             left: 0,
             top: 0,